app/requirements.txt
notion-haystack
haystack-ai
sentence-transformers~=3.3.1
openai~=1.59.3
uvicorn~=0.34.0
fastapi~=0.115.6
pydantic~=2.10.4
notion-client~=2.3.0
python-dotenv~=1.0.1
haystack~=0.42
slack_sdk~=3.34.0
Flask~=3.1.0
SQLAlchemy~=2.0.36
tiktoken~=0.8.0
preprocessor~=1.1.3
qdrant-client~=1.12.2
Flask-Cors~=5.0.0
slack_bolt~=1.22.0
qdrant-haystack